requirements.txt
python-binance
TA-Lib
numpy
websocket_client
python-telegram-bot
colorama
ccxt 
schedule
pandas
schedule